backend 2nd

+++ novatech_be/src/main/java/com/nvl/novatech/service/ReviewServiceImpl.java
@@ -1,10 +1,12 @@
 package com.nvl.novatech.service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
 
 import com.nvl.novatech.dto.request.ReviewRequest;
+import com.nvl.novatech.dto.request.ReviewUpdateRequest;
 import com.nvl.novatech.exception.AppException;
 import com.nvl.novatech.exception.ErrorCode;
 import com.nvl.novatech.model.Product;
@@ -34,25 +36,29 @@ public class ReviewServiceImpl implements ReviewService{
         review.setRating(request.getRating());
         review.setProduct(product);
         review.setUser(user);
+        review.setCreateAt(LocalDateTime.now());
         return reviewRepository.save(review);
     }
 
     @Override
-    public Review updateReview(Long reviewId, User user, ReviewRequest request) {
-        Product product = productRepository.findById(request.getProductId()).orElseThrow(() -> new AppException(ErrorCode.PRODUCT_NOT_EXISTED));
+    public Review updateReview(Long reviewId, User user, ReviewUpdateRequest request) {
         Review review = reviewRepository.findById(reviewId).orElseThrow(() -> new AppException(ErrorCode.REVIEW_NOT_EXISTED));
         review.setContent(request.getContent());
         review.setImageUrl(request.getImageUrl());
-        request.setRating(request.getRating());
-        review.setProduct(product);
-        review.setUser(user);
+        review.setRating(request.getRating());
+        review.setCreateAt(LocalDateTime.now());
         return reviewRepository.save(review);
     }
 
     @Override
     public void deleteReview(Long reviewId, User user) {
-        Review review = reviewRepository.findReviewByUserId(user.getUserId());
-        reviewRepository.delete(review);
+        List<Review> listReview = reviewRepository.findReviewByUserId(user.getUserId());
+        Review rv = reviewRepository.findById(reviewId).orElseThrow(() -> new AppException(ErrorCode.REVIEW_NOT_EXISTED));
+        for (Review review : listReview) {
+            if(review.getReviewId().equals(rv.getReviewId())){
+                reviewRepository.delete(review);
+            }
+        }
     }
 
     @Override
